policies + tag + fin catégorie

+++ app/Http/Controllers/ArticleController.php
@@ -6,6 +6,7 @@ use Illuminate\Http\Request;
 use App\Http\Requests\StoreArticle;
 use App\Article;
 use App\Category;
+use Auth;
 
 class ArticleController extends Controller
 {
@@ -25,19 +26,31 @@ class ArticleController extends Controller
     /* -- ESPACE ADMINISTRATION -- */
     // View
     public function list() {
-        $articles = Article::get();
-        return view('admin.article.list', compact('articles'));
+        if (Auth::user()->can('list', Article::class)) {
+            $articles = Article::get();
+            return view('admin.article.list', compact('articles'));
+        } else {
+            return redirect(route('admin.dashboard'));
+        }
     }
 
     public function edit($id) {
         $article = Article::findOrFail($id);   
-        $categories = Category::get(['name', 'id']);
-        return view('admin.article.edit', compact('article', 'categories'));
+        if (Auth::user()->can('update', $article)) {
+            $categories = Category::get(['name', 'id']);
+            return view('admin.article.edit', compact('article', 'categories'));
+        } else {
+            return redirect(route('admin.articles'));
+        }
     }
 
     public function add() {
-        $categories = Category::get(['name', 'id']);
-        return view('admin.article.add', compact('categories'));
+        if (Auth::user()->can('create', Article::class)) {
+            $categories = Category::get(['name', 'id']);
+            return view('admin.article.add', compact('categories'));
+        } else {
+            return redirect(route('admin.articles'));
+        }
     }
 
     /*
@@ -45,52 +58,58 @@ class ArticleController extends Controller
     */
     public function update($id, StoreArticle $request) {
         $article = Article::findOrFail($id);
-        $article->content = $request->content;
-        $article->title = $request->title;
+        if (Auth::user()->can('update', $article)) {
+            $article->content = $request->content;
+            $article->title = $request->title;
 
-        // Slug
-        $article->slug = str_replace( array(' ', '\'', '"'), array('_', '', ''), $request->slug ? $request->slug : $request->title );
+            // Slug
+            $article->slug = str_replace( array(' ', '\'', '"'), array('_', '', ''), $request->slug ? $request->slug : $request->title );
 
-        // Picture
-        if( $request->hasFile('picture') ) {
-            $file = $request->file('picture');
-            $name = 'article_' . $id . '_' . $file->getClientOriginalName();
-            $file->move($this->storage.'/'. date('Y') , $name);
-            $article->picture = $this->storage .'/'. date('Y') .'/'. $name;
+            // Picture
+            if( $request->hasFile('picture') ) {
+                $file = $request->file('picture');
+                $name = 'article_' . $id . '_' . $file->getClientOriginalName();
+                $file->move($this->storage.'/'. date('Y') , $name);
+                $article->picture = $this->storage .'/'. date('Y') .'/'. $name;
+            }
+            $article->save();
+
+            // Category
+            $category = Category::where('id', $request->category)->first();
+            $article->categories()->attach($category);
+
+            return redirect( route('admin.article.edit', compact('article')) )->withSuccess('Article Modifié');
+        } else {
+            return redirect(route('admin.articles'));
         }
-        $article->save();
-
-        // Category
-        $category = Category::where('id', $request->category)->first();
-        $article->categories()->attach($category);
-
-        return redirect( route('admin.article.edit', compact('article')) )->withSuccess('Article Modifié');
     }
     
     /*
     *   Store new Article
     */
     public function store(StoreArticle $request) {
-        $article = new Article;
-        $article->draft = 0;
-        $article->title = $request->title;
-        $article->content = $request->content;
+        if (Auth::user()->can('create', Article::class)) {
+            $article = new Article;
+            $article->draft = 1;
+            $article->title = $request->title;
+            $article->content = $request->content;
 
-        // Slug
-        $article->slug = strtolower( str_replace( array(' ', '_', '\'', '"'), array('-','-', '', ''), $request->slug ? $request->slug : $request->title ) );
-    
-        // Picture [ rework name : _{id}_ actually missing ]
-        if( $request->hasFile('picture') ) {
-            $file = $request->file('picture');
-            $name = 'article_' . $file->getClientOriginalName();
-            $file->move($this->storage .'/'. date('Y'), $name);
-            $article->picture = $this->storage .'/'. date('Y') .'/'. $name;
+            // Slug
+            $article->slug = strtolower( str_replace( array(' ', '_', '\'', '"'), array('-','-', '', ''), $request->slug ? $request->slug : $request->title ) );
+        
+            // Picture [ rework name : _{id}_ actually missing ]
+            if( $request->hasFile('picture') ) {
+                $file = $request->file('picture');
+                $name = 'article_' . $file->getClientOriginalName();
+                $file->move($this->storage .'/'. date('Y'), $name);
+                $article->picture = $this->storage .'/'. date('Y') .'/'. $name;
+            }
+            $article->save();
+
+            // Category 
+            $category = Category::where('id', $request->category)->first();
+            $article->categories()->attach($category);
         }
-        $article->save();
-
-        // Category 
-        $category = Category::where('id', $request->category)->first();
-        $article->categories()->attach($category);
 
         return redirect( route('admin.articles') );
     }
@@ -99,8 +118,10 @@ class ArticleController extends Controller
     public function delete(Request $request) {
         $id = $request->input('id');
         $article = Article::findOrFail($id);
-        if( $article->delete() ){
-            return response()->json(['status' => true]);
+        if (Auth::user()->can('delete', $article)) {
+            if( $article->delete() ){
+                return response()->json(['status' => true]);
+            }
         }
         return response()->json(['status' => false]);
     }
@@ -112,9 +133,11 @@ class ArticleController extends Controller
     public function unpublish(Request $request) {
         $id = $request->input('id');
         $article = Article::findOrFail($id);
-        $article->draft = 1;
-        if( $article->save() ) {
-            return response()->json(['status' => true]);
+        if (Auth::user()->can('unpublish', $article)) {
+            $article->draft = 1;
+            if( $article->save() ) {
+                return response()->json(['status' => true]);
+            }
         }
         return response()->json(['status' => false]);
     }
@@ -122,9 +145,11 @@ class ArticleController extends Controller
     public function publish(Request $request) {
         $id = $request->input('id');
         $article = Article::findOrFail($id);
-        $article->draft = 0;
-        if( $article->save() ) {
-            return response()->json(['status' => true]);
+        if (Auth::user()->can('publish', $article)) {
+            $article->draft = 0;
+            if( $article->save() ) {
+                return response()->json(['status' => true]);
+            }
         }
         return response()->json(['status' => false]);
     }
