+++ app/Http/Controllers/CategoryController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Http\Requests\StoreCategory;
 use App\Category;
+use Auth;
 
 class CategoryController extends Controller
 {
@@ -15,30 +16,48 @@ class CategoryController extends Controller
 
     /* -- ESPACE ADMINISTRATION -- */
     public function list() {
-        $categories = Category::get();
-        return view('admin.category.list', compact('categories'));
+        if (Auth::user()->can('list', Category::class)) {
+            $categories = Category::get();
+            return view('admin.category.list', compact('categories'));
+        }
+        return redirect(route('admin.dashboard'));
     }
 
-    public function edit() {
-        return view('admin.category.edit');
+    public function edit($id) {
+        $category = Category::findOrFail($id);   
+        if (Auth::user()->can('update', $category)) {
+            return view('admin.category.edit', compact('category'));
+        }
+        return redirect(route('admin.categories'));
     }
 
     public function add() {
-        return view('admin.category.add');
+        if (Auth::user()->can('create', Category::class)) {
+            return view('admin.category.add');
+        }
+        return redirect(route('admin.categories'));
     }
 
-    public function update() {
+    public function update($id, StoreCategory $request) {
+        $category = Category::findOrFail($id);
+        if (Auth::user()->can('update', $category)) {
+            $category->name = $request->name;
+            $category->slug = $request->slug;
+            $category->description = $request->description;
+            $category->save();
+            return redirect( route('admin.category.edit', compact('id')) )->withSuccess('Catégorie Modifiée');
+        }
         return redirect(route('admin.categories'));
     }
 
     public function store(StoreCategory $request) {
-
-        $category = new Category;
-        $category->name = $request->name;
-        $category->slug = $request->slug;
-        $category->description = $request->description;
-        $category->save();
-
+        if (Auth::user()->can('create', Category::class)) {
+            $category = new Category;
+            $category->name = $request->name;
+            $category->slug = $request->slug;
+            $category->description = $request->description;
+            $category->save();
+        }
         return redirect(route('admin.categories'));
     }
 
@@ -46,8 +65,10 @@ class CategoryController extends Controller
     public function delete(Request $request) {
         $id = $request->input('id');
         $category = Category::findOrFail($id);
-        if( $category->delete() ){
-            return response()->json(['status' => true]);
+        if (Auth::user()->can('delete', $category)) {
+            if( $category->delete() ){
+                return response()->json(['status' => true]);
+            }
         }
         return response()->json(['status' => false]);
     }
